refactor: restructure default components
