+++ src/index.js
@@ -1,0 +1,3 @@
+export { Button } from './components/Button/Button'
+export { Header } from './components/Header/Header'
+export { Page } from './components/Page/Page'
